voice-assistant/requirements.txt
python-dotenv
bring-api
mashumaro
orjson
picovoice
pvporcupine
requests
rich
sounddevice
scipy
pvrecorder
aiohttp
piper-tts
gTTS